projects-templates/k8s_v1/scripts/pyproject.toml
[tool.poetry]
name = "lib"
version = "1.0.0"
description = "Scripts for k8s_v1"
authors = ["dantas74 <m47heusdr@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
typer = "^0.9.0"
colorama = "^0.4.6"
shellingham = "^1.5.4"
PyYAML = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
